Migrate casts from int literals to use double and float literals
RELNOTES=n/a
PiperOrigin-RevId: 746063845

+++ android/guava-tests/test/com/google/common/primitives/DoubleArrayAsListTest.java
@@ -101,7 +101,7 @@ public class DoubleArrayAsListTest extends TestCase {
   public static final class DoublesAsListTailSubListGenerator extends TestDoubleListGenerator {
     @Override
     protected List<Double> create(Double[] elements) {
-      Double[] prefix = {(double) 86, (double) 99};
+      Double[] prefix = {86.0, 99.0};
       Double[] all = concat(prefix, elements);
       return asList(all).subList(2, elements.length + 2);
     }
@@ -111,7 +111,7 @@ public class DoubleArrayAsListTest extends TestCase {
     @Override
     protected List<Double> create(Double[] elements) {
       Double[] prefix = {Double.MIN_VALUE, Double.MAX_VALUE};
-      Double[] suffix = {(double) 86, (double) 99};
+      Double[] suffix = {86.0, 99.0};
       Double[] all = concat(concat(prefix, elements), suffix);
       return asList(all).subList(2, elements.length + 2);
     }
@@ -160,7 +160,7 @@ public class DoubleArrayAsListTest extends TestCase {
 
   public static class SampleDoubles extends SampleElements<Double> {
     public SampleDoubles() {
-      super((double) 0, (double) 1, (double) 2, (double) 3, (double) 4);
+      super(0.0, 1.0, 2.0, 3.0, 4.0);
     }
   }
 }
