+++ android/guava-tests/test/com/google/common/primitives/DoublesTest.java
@@ -50,8 +50,8 @@ import org.jspecify.annotations.Nullable;
 @GwtCompatible(emulated = true)
 public class DoublesTest extends TestCase {
   private static final double[] EMPTY = {};
-  private static final double[] ARRAY1 = {(double) 1};
-  private static final double[] ARRAY234 = {(double) 2, (double) 3, (double) 4};
+  private static final double[] ARRAY1 = {1.0};
+  private static final double[] ARRAY234 = {2.0, 3.0, 4.0};
 
   private static final double LEAST = Double.NEGATIVE_INFINITY;
   private static final double GREATEST = Double.POSITIVE_INFINITY;
@@ -109,13 +109,13 @@ public class DoublesTest extends TestCase {
   }
 
   public void testContains() {
-    assertThat(Doubles.contains(EMPTY, (double) 1)).isFalse();
-    assertThat(Doubles.contains(ARRAY1, (double) 2)).isFalse();
-    assertThat(Doubles.contains(ARRAY234, (double) 1)).isFalse();
-    assertThat(Doubles.contains(new double[] {(double) -1}, (double) -1)).isTrue();
-    assertThat(Doubles.contains(ARRAY234, (double) 2)).isTrue();
-    assertThat(Doubles.contains(ARRAY234, (double) 3)).isTrue();
-    assertThat(Doubles.contains(ARRAY234, (double) 4)).isTrue();
+    assertThat(Doubles.contains(EMPTY, 1.0)).isFalse();
+    assertThat(Doubles.contains(ARRAY1, 2.0)).isFalse();
+    assertThat(Doubles.contains(ARRAY234, 1.0)).isFalse();
+    assertThat(Doubles.contains(new double[] {-1.0}, -1.0)).isTrue();
+    assertThat(Doubles.contains(ARRAY234, 2.0)).isTrue();
+    assertThat(Doubles.contains(ARRAY234, 3.0)).isTrue();
+    assertThat(Doubles.contains(ARRAY234, 4.0)).isTrue();
 
     for (double value : NUMBERS) {
       assertWithMessage("" + value)
@@ -126,17 +126,14 @@ public class DoublesTest extends TestCase {
   }
 
   public void testIndexOf() {
-    assertThat(Doubles.indexOf(EMPTY, (double) 1)).isEqualTo(-1);
-    assertThat(Doubles.indexOf(ARRAY1, (double) 2)).isEqualTo(-1);
-    assertThat(Doubles.indexOf(ARRAY234, (double) 1)).isEqualTo(-1);
-    assertThat(Doubles.indexOf(new double[] {(double) -1}, (double) -1)).isEqualTo(0);
-    assertThat(Doubles.indexOf(ARRAY234, (double) 2)).isEqualTo(0);
-    assertThat(Doubles.indexOf(ARRAY234, (double) 3)).isEqualTo(1);
-    assertThat(Doubles.indexOf(ARRAY234, (double) 4)).isEqualTo(2);
-    assertThat(
-            Doubles.indexOf(
-                new double[] {(double) 2, (double) 3, (double) 2, (double) 3}, (double) 3))
-        .isEqualTo(1);
+    assertThat(Doubles.indexOf(EMPTY, 1.0)).isEqualTo(-1);
+    assertThat(Doubles.indexOf(ARRAY1, 2.0)).isEqualTo(-1);
+    assertThat(Doubles.indexOf(ARRAY234, 1.0)).isEqualTo(-1);
+    assertThat(Doubles.indexOf(new double[] {-1.0}, -1.0)).isEqualTo(0);
+    assertThat(Doubles.indexOf(ARRAY234, 2.0)).isEqualTo(0);
+    assertThat(Doubles.indexOf(ARRAY234, 3.0)).isEqualTo(1);
+    assertThat(Doubles.indexOf(ARRAY234, 4.0)).isEqualTo(2);
+    assertThat(Doubles.indexOf(new double[] {2.0, 3.0, 2.0, 3.0}, 3.0)).isEqualTo(1);
 
     for (double value : NUMBERS) {
       assertWithMessage("" + value)
@@ -154,33 +151,21 @@ public class DoublesTest extends TestCase {
     assertThat(Doubles.indexOf(ARRAY1, ARRAY234)).isEqualTo(-1);
     assertThat(Doubles.indexOf(ARRAY1, ARRAY1)).isEqualTo(0);
     assertThat(Doubles.indexOf(ARRAY234, ARRAY234)).isEqualTo(0);
-    assertThat(Doubles.indexOf(ARRAY234, new double[] {(double) 2, (double) 3})).isEqualTo(0);
-    assertThat(Doubles.indexOf(ARRAY234, new double[] {(double) 3, (double) 4})).isEqualTo(1);
-    assertThat(Doubles.indexOf(ARRAY234, new double[] {(double) 3})).isEqualTo(1);
-    assertThat(Doubles.indexOf(ARRAY234, new double[] {(double) 4})).isEqualTo(2);
-    assertThat(
-            Doubles.indexOf(
-                new double[] {(double) 2, (double) 3, (double) 3, (double) 3, (double) 3},
-                new double[] {(double) 3}))
+    assertThat(Doubles.indexOf(ARRAY234, new double[] {2.0, 3.0})).isEqualTo(0);
+    assertThat(Doubles.indexOf(ARRAY234, new double[] {3.0, 4.0})).isEqualTo(1);
+    assertThat(Doubles.indexOf(ARRAY234, new double[] {3.0})).isEqualTo(1);
+    assertThat(Doubles.indexOf(ARRAY234, new double[] {4.0})).isEqualTo(2);
+    assertThat(Doubles.indexOf(new double[] {2.0, 3.0, 3.0, 3.0, 3.0}, new double[] {3.0}))
         .isEqualTo(1);
     assertThat(
             Doubles.indexOf(
-                new double[] {
-                  (double) 2, (double) 3, (double) 2, (double) 3, (double) 4, (double) 2, (double) 3
-                },
-                new double[] {(double) 2, (double) 3, (double) 4}))
+                new double[] {2.0, 3.0, 2.0, 3.0, 4.0, 2.0, 3.0}, new double[] {2.0, 3.0, 4.0}))
         .isEqualTo(2);
     assertThat(
             Doubles.indexOf(
-                new double[] {
-                  (double) 2, (double) 2, (double) 3, (double) 4, (double) 2, (double) 3, (double) 4
-                },
-                new double[] {(double) 2, (double) 3, (double) 4}))
+                new double[] {2.0, 2.0, 3.0, 4.0, 2.0, 3.0, 4.0}, new double[] {2.0, 3.0, 4.0}))
         .isEqualTo(1);
-    assertThat(
-            Doubles.indexOf(
-                new double[] {(double) 4, (double) 3, (double) 2},
-                new double[] {(double) 2, (double) 3, (double) 4}))
+    assertThat(Doubles.indexOf(new double[] {4.0, 3.0, 2.0}, new double[] {2.0, 3.0, 4.0}))
         .isEqualTo(-1);
 
     for (double value : NUMBERS) {
@@ -193,17 +178,14 @@ public class DoublesTest extends TestCase {
   }
 
   public void testLastIndexOf() {
-    assertThat(Doubles.lastIndexOf(EMPTY, (double) 1)).isEqualTo(-1);
-    assertThat(Doubles.lastIndexOf(ARRAY1, (double) 2)).isEqualTo(-1);
-    assertThat(Doubles.lastIndexOf(ARRAY234, (double) 1)).isEqualTo(-1);
-    assertThat(Doubles.lastIndexOf(new double[] {(double) -1}, (double) -1)).isEqualTo(0);
-    assertThat(Doubles.lastIndexOf(ARRAY234, (double) 2)).isEqualTo(0);
-    assertThat(Doubles.lastIndexOf(ARRAY234, (double) 3)).isEqualTo(1);
-    assertThat(Doubles.lastIndexOf(ARRAY234, (double) 4)).isEqualTo(2);
-    assertThat(
-            Doubles.lastIndexOf(
-                new double[] {(double) 2, (double) 3, (double) 2, (double) 3}, (double) 3))
-        .isEqualTo(3);
+    assertThat(Doubles.lastIndexOf(EMPTY, 1.0)).isEqualTo(-1);
+    assertThat(Doubles.lastIndexOf(ARRAY1, 2.0)).isEqualTo(-1);
+    assertThat(Doubles.lastIndexOf(ARRAY234, 1.0)).isEqualTo(-1);
+    assertThat(Doubles.lastIndexOf(new double[] {-1.0}, -1.0)).isEqualTo(0);
+    assertThat(Doubles.lastIndexOf(ARRAY234, 2.0)).isEqualTo(0);
+    assertThat(Doubles.lastIndexOf(ARRAY234, 3.0)).isEqualTo(1);
+    assertThat(Doubles.lastIndexOf(ARRAY234, 4.0)).isEqualTo(2);
+    assertThat(Doubles.lastIndexOf(new double[] {2.0, 3.0, 2.0, 3.0}, 3.0)).isEqualTo(3);
 
     for (double value : NUMBERS) {
       assertWithMessage("" + value)
@@ -221,9 +203,7 @@ public class DoublesTest extends TestCase {
   public void testMax() {
     assertThat(max(LEAST)).isEqualTo(LEAST);
     assertThat(max(GREATEST)).isEqualTo(GREATEST);
-    assertThat(
-            max((double) 8, (double) 6, (double) 7, (double) 5, (double) 3, (double) 0, (double) 9))
-        .isEqualTo((double) 9);
+    assertThat(max(8.0, 6.0, 7.0, 5.0, 3.0, 0.0, 9.0)).isEqualTo(9.0);
 
     assertThat(max(-0.0, 0.0)).isEqualTo(0.0);
     assertThat(max(0.0, -0.0)).isEqualTo(0.0);
@@ -239,9 +219,7 @@ public class DoublesTest extends TestCase {
   public void testMin() {
     assertThat(min(LEAST)).isEqualTo(LEAST);
     assertThat(min(GREATEST)).isEqualTo(GREATEST);
-    assertThat(
-            min((double) 8, (double) 6, (double) 7, (double) 5, (double) 3, (double) 0, (double) 9))
-        .isEqualTo((double) 0);
+    assertThat(min(8.0, 6.0, 7.0, 5.0, 3.0, 0.0, 9.0)).isEqualTo(0.0);
 
     assertThat(min(-0.0, 0.0)).isEqualTo(-0.0);
     assertThat(min(0.0, -0.0)).isEqualTo(-0.0);
@@ -250,15 +228,12 @@ public class DoublesTest extends TestCase {
   }
 
   public void testConstrainToRange() {
-    assertThat(Doubles.constrainToRange((double) 1, (double) 0, (double) 5)).isEqualTo((double) 1);
-    assertThat(Doubles.constrainToRange((double) 1, (double) 1, (double) 5)).isEqualTo((double) 1);
-    assertThat(Doubles.constrainToRange((double) 1, (double) 3, (double) 5)).isEqualTo((double) 3);
-    assertThat(Doubles.constrainToRange((double) 0, (double) -5, (double) -1))
-        .isEqualTo((double) -1);
-    assertThat(Doubles.constrainToRange((double) 5, (double) 2, (double) 2)).isEqualTo((double) 2);
-    assertThrows(
-        IllegalArgumentException.class,
-        () -> Doubles.constrainToRange((double) 1, (double) 3, (double) 2));
+    assertThat(Doubles.constrainToRange(1.0, 0.0, 5.0)).isEqualTo(1.0);
+    assertThat(Doubles.constrainToRange(1.0, 1.0, 5.0)).isEqualTo(1.0);
+    assertThat(Doubles.constrainToRange(1.0, 3.0, 5.0)).isEqualTo(3.0);
+    assertThat(Doubles.constrainToRange(0.0, -5.0, -1.0)).isEqualTo(-1.0);
+    assertThat(Doubles.constrainToRange(5.0, 2.0, 2.0)).isEqualTo(2.0);
+    assertThrows(IllegalArgumentException.class, () -> Doubles.constrainToRange(1.0, 3.0, 2.0));
   }
 
   public void testConcat() {
@@ -268,10 +243,8 @@ public class DoublesTest extends TestCase {
     assertThat(Doubles.concat(ARRAY1)).isEqualTo(ARRAY1);
     assertThat(Doubles.concat(ARRAY1)).isNotSameInstanceAs(ARRAY1);
     assertThat(Doubles.concat(EMPTY, ARRAY1, EMPTY)).isEqualTo(ARRAY1);
-    assertThat(Doubles.concat(ARRAY1, ARRAY1, ARRAY1))
-        .isEqualTo(new double[] {(double) 1, (double) 1, (double) 1});
-    assertThat(Doubles.concat(ARRAY1, ARRAY234))
-        .isEqualTo(new double[] {(double) 1, (double) 2, (double) 3, (double) 4});
+    assertThat(Doubles.concat(ARRAY1, ARRAY1, ARRAY1)).isEqualTo(new double[] {1.0, 1.0, 1.0});
+    assertThat(Doubles.concat(ARRAY1, ARRAY234)).isEqualTo(new double[] {1.0, 2.0, 3.0, 4.0});
   }
 
   @GwtIncompatible // different overflow behavior; could probably be made to work by using ~~
@@ -309,10 +282,7 @@ public class DoublesTest extends TestCase {
     assertThat(Doubles.ensureCapacity(EMPTY, 0, 1)).isSameInstanceAs(EMPTY);
     assertThat(Doubles.ensureCapacity(ARRAY1, 0, 1)).isSameInstanceAs(ARRAY1);
     assertThat(Doubles.ensureCapacity(ARRAY1, 1, 1)).isSameInstanceAs(ARRAY1);
-    assertThat(
-            Arrays.equals(
-                new double[] {(double) 1, (double) 0, (double) 0},
-                Doubles.ensureCapacity(ARRAY1, 2, 1)))
+    assertThat(Arrays.equals(new double[] {1.0, 0.0, 0.0}, Doubles.ensureCapacity(ARRAY1, 2, 1)))
         .isTrue();
   }
 
@@ -325,8 +295,8 @@ public class DoublesTest extends TestCase {
   public void testJoin() {
     assertThat(Doubles.join(",", EMPTY)).isEmpty();
     assertThat(Doubles.join(",", ARRAY1)).isEqualTo("1.0");
-    assertThat(Doubles.join(",", (double) 1, (double) 2)).isEqualTo("1.0,2.0");
-    assertThat(Doubles.join("", (double) 1, (double) 2, (double) 3)).isEqualTo("1.02.03.0");
+    assertThat(Doubles.join(",", 1.0, 2.0)).isEqualTo("1.0,2.0");
+    assertThat(Doubles.join("", 1.0, 2.0, 3.0)).isEqualTo("1.02.03.0");
   }
 
   public void testJoinNonTrivialDoubles() {
@@ -342,9 +312,9 @@ public class DoublesTest extends TestCase {
             new double[] {},
             new double[] {LEAST},
             new double[] {LEAST, LEAST},
-            new double[] {LEAST, (double) 1},
-            new double[] {(double) 1},
-            new double[] {(double) 1, LEAST},
+            new double[] {LEAST, 1.0},
+            new double[] {1.0},
+            new double[] {1.0, LEAST},
             new double[] {GREATEST, Double.MAX_VALUE},
             new double[] {GREATEST, GREATEST},
             new double[] {GREATEST, GREATEST, GREATEST});
@@ -537,12 +507,12 @@ public class DoublesTest extends TestCase {
     List<Double> none = Arrays.<Double>asList();
     assertThat(Doubles.toArray(none)).isEqualTo(EMPTY);
 
-    List<Double> one = Arrays.asList((double) 1);
+    List<Double> one = Arrays.asList(1.0);
     assertThat(Doubles.toArray(one)).isEqualTo(ARRAY1);
 
-    double[] array = {(double) 0, (double) 1, Math.PI};
+    double[] array = {0.0, 1.0, Math.PI};
 
-    List<Double> three = Arrays.asList((double) 0, (double) 1, Math.PI);
+    List<Double> three = Arrays.asList(0.0, 1.0, Math.PI);
     assertThat(Doubles.toArray(three)).isEqualTo(array);
 
     assertThat(Doubles.toArray(Doubles.asList(array))).isEqualTo(array);
@@ -564,19 +534,19 @@ public class DoublesTest extends TestCase {
   }
 
   public void testToArray_withNull() {
-    List<@Nullable Double> list = Arrays.asList((double) 0, (double) 1, null);
+    List<@Nullable Double> list = Arrays.asList(0.0, 1.0, null);
     assertThrows(NullPointerException.class, () -> Doubles.toArray(list));
   }
 
   public void testToArray_withConversion() {
-    double[] array = {(double) 0, (double) 1, (double) 2};
+    double[] array = {0.0, 1.0, 2.0};
 
     List<Byte> bytes = Arrays.asList((byte) 0, (byte) 1, (byte) 2);
     List<Short> shorts = Arrays.asList((short) 0, (short) 1, (short) 2);
     List<Integer> ints = Arrays.asList(0, 1, 2);
-    List<Float> floats = Arrays.asList((float) 0, (float) 1, (float) 2);
+    List<Float> floats = Arrays.asList(0.0f, 1.0f, 2.0f);
     List<Long> longs = Arrays.asList(0L, 1L, 2L);
-    List<Double> doubles = Arrays.asList((double) 0, (double) 1, (double) 2);
+    List<Double> doubles = Arrays.asList(0.0, 1.0, 2.0);
 
     assertThat(Doubles.toArray(bytes)).isEqualTo(array);
     assertThat(Doubles.toArray(shorts)).isEqualTo(array);
@@ -588,32 +558,31 @@ public class DoublesTest extends TestCase {
 
   @J2ktIncompatible // b/239034072: Kotlin varargs copy parameter arrays.
   public void testAsList_isAView() {
-    double[] array = {(double) 0, (double) 1};
+    double[] array = {0.0, 1.0};
     List<Double> list = Doubles.asList(array);
-    list.set(0, (double) 2);
-    assertThat(array).isEqualTo(new double[] {(double) 2, (double) 1});
-    array[1] = (double) 3;
-    assertThat(list).containsExactly((double) 2, (double) 3).inOrder();
+    list.set(0, 2.0);
+    assertThat(array).isEqualTo(new double[] {2.0, 1.0});
+    array[1] = 3.0;
+    assertThat(list).containsExactly(2.0, 3.0).inOrder();
   }
 
   public void testAsList_toArray_roundTrip() {
-    double[] array = {(double) 0, (double) 1, (double) 2};
+    double[] array = {0.0, 1.0, 2.0};
     List<Double> list = Doubles.asList(array);
     double[] newArray = Doubles.toArray(list);
 
     // Make sure it returned a copy
-    list.set(0, (double) 4);
-    assertThat(newArray).isEqualTo(new double[] {(double) 0, (double) 1, (double) 2});
-    newArray[1] = (double) 5;
-    assertThat((double) list.get(1)).isEqualTo((double) 1);
+    list.set(0, 4.0);
+    assertThat(newArray).isEqualTo(new double[] {0.0, 1.0, 2.0});
+    newArray[1] = 5.0;
+    assertThat((double) list.get(1)).isEqualTo(1.0);
   }
 
   // This test stems from a real bug found by andrewk
   public void testAsList_subList_toArray_roundTrip() {
-    double[] array = {(double) 0, (double) 1, (double) 2, (double) 3};
+    double[] array = {0.0, 1.0, 2.0, 3.0};
     List<Double> list = Doubles.asList(array);
-    assertThat(Doubles.toArray(list.subList(1, 3)))
-        .isEqualTo(new double[] {(double) 1, (double) 2});
+    assertThat(Doubles.toArray(list.subList(1, 3))).isEqualTo(new double[] {1.0, 2.0});
     assertThat(Doubles.toArray(list.subList(2, 2))).isEmpty();
   }
